+++ priv/repo/migrations/20181006004707_add_base_hp_to_monsters.exs
@@ -1,0 +1,9 @@
+defmodule ApathyDrive.Repo.Migrations.AddBaseHpToMonsters do
+  use Ecto.Migration
+
+  def change do
+    alter table(:monsters) do
+      add(:base_hp, :integer)
+    end
+  end
+end